interface edited

+++ Spice.cpp
@@ -4,5 +4,5 @@ Spice::Spice(const std::string& name, int basePrice, int amount, double purity)
     : Cargo(name, basePrice, amount), purity_(purity) {}
 
 int Spice::getPrice() const {
-    return basePrice_ * purity_ / bestPurity;
+    return static_cast<int>(basePrice_ * purity_ / bestPurity);
 }
